Backend now steps, adding logo assets, and CircleCI logo
After fixing the CircleCI integration, I created an AltHack logo and
added it into the new `assets` folder. Also, backends are supposed to
be able to step when the main loop steps, so that was added (empty for
now).

The README.rst file now reflects the current CircleCI build status.