Adjusting main window layout and adding grid
The main window layout hasn't converged yet fully, and this
commit makes adjustments to it to make it more appealing.

Also, a grid is now drawn on the central canvas. This grid can
be customized in column/row appearance, and will act as both a
visual and a programmatic guide for placing nodes. The canvas
can be dragged around, and the grid moves accordingly. Drawing
needs to be adjusted to draw the origin properly. but in general
the logic for canvas drag is implemented.