Fixed canvas dragging and grid drawing
The drawing code for the canvas when dragged wasn't handling drawing
outside of bunds properly. The canvas can now be dragged around freely,
and only the relevant part of the grid in the background is drawn.

Some minor parts of the code were cleaned up, and a text output for
the current canvas position was added.