Communication and scheduling for backends
Backends need to be wired up properly to allow data storage, retrieval,
and to decouple visualization from game logic. To that end, they need
their own scheduling cycle, and a mechanism to start and end it that is
in sync with the main application thread.

This change adds that mechanism and ensures that backends do not run in
the same thread as the main visualization/game logic. Also, the account
data used in the game now comes from the backend initially rather than
being hardcoded in the frontend.