Move position variable reference info canvas method
Since the canvas position member variable of MainWindow is updated
inside `canvas()`, reading it for the purpose of populating the canvas
widget should also happen inside that method.

This change removes the parameter of `canvas()` for the current canvas
position, and lets the method read the position from the member variable
directly.

Closes issue #2.