Integrate database into server backend
To be able to store data persistently, a database storage layer needs to
be integrated into the server backend.

This change integrates SOCI as a transparent layer to SQLITE3. The
reason for this choice is that SOCI adds modern C++ mechanisms to the
file-based approach of SQLITE, which is both very much desirable for the
project.

Relates to issue #19.