Creating actual database file and creating required tables on start
To make the database usable, it needs to be populated with initial
tables and data when creating it. This change adds the relevant
interfaces for intentionally opening and closing the database interface,
adding tables, and adding an example user.

The exact data to be fed into the database still needs to be decided,
but the database backend should now be properly integrated, ready for
storing persistent data. More interfaces to add are subject to exact
use-cases.