Add a frontend layer to communicate with the backend
The frontend layer can connect to either a local server backend, or to a
future proxy backend that connects to a remote server backend. The
backend interface dictates the available functionality for the frontend
instance.

The frontend requires a worker thread on its own, which communicates
with the backend, updates the UI, but does not interfere with the UI
thread OR the backend thread.