Added database functions for storage and management of data
The SOCI database backend as well as the general database interface got
functions for storing and managing data. This includes creation of
arbitrary entities, adding properties to them, and deleting either of
them again. The only thing missing in this one is the retrieval part,
which will be done next.